+++ src/rybak/_types.py
@@ -1,3 +1,4 @@
+import pathlib
 from typing import Any, Callable, Iterable, Mapping, TypeVar
 
 from .pycompat import TypeAlias
@@ -7,6 +8,8 @@ TemplateData: TypeAlias = Mapping[str, Any]
 Item = TypeVar('Item')
 LoopOverFn: TypeAlias = Callable[[Iterable[Item]], Item]
 RenderFn: TypeAlias = Callable[[str, str, TemplateData], None]
+
+ReportCallbackFn: TypeAlias = Callable[[pathlib.PurePath, pathlib.Path], None]
 
 
 class RenderError(Exception):