+++ CHANGELOG.md
@@ -1,6 +1,6 @@
 # Changelog
 
-# [Unreleased]
+# 0.3.0 - 2024-02-24
 
 ### Changed